+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 1.2.0
+
+- Introduced `min_version` in metadata for indicating that a certain version of `semgrep-search` is required for the database
+- Moved repositories to configuration file
+- Added licenses for repositories
+- Added more repositories
+
 ## Version 1.1.0
 
 - Normalized language in index data to their base name